chore: update loki version, refactor fields functions to service
