+++ package.json
@@ -11,6 +11,7 @@
     "lint": "eslint --cache --ignore-path ./.gitignore --ext .js,.jsx,.ts,.tsx .",
     "lint:fix": "npm run lint -- --fix",
     "e2e": "playwright test -x --trace on",
+    "e2e:fast": "playwright test --retries=1 --workers=3",
     "e2e:10x": "playwright test -x --trace on --repeat-each=10 --workers 1",
     "e2e:10x:fast": "playwright test --repeat-each=10 --workers 3 --retries=1",
     "e2e:flake": "yarn e2e:10x",
